+++ src/python/pants/backend/experimental/go/lint/gofmt/register.py
@@ -1,0 +1,9 @@
+# Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+from pants.backend.go.lint import fmt
+from pants.backend.go.lint.gofmt.rules import rules as gofmt_rules
+
+
+def rules():
+    return [*fmt.rules(), *gofmt_rules()]
