+++ src/python/pants/backend/go/lint/fmt.py
