+++ src/python/pants/backend/go/lint/gofmt/rules_integration_test.py
@@ -5,8 +5,9 @@ from typing import List, Sequence, Tuple
 
 import pytest
 
-from pants.backend.go import fmt, gofmt
 from pants.backend.go.gofmt import GofmtFieldSet, GofmtRequest
+from pants.backend.go.lint import fmt
+from pants.backend.go.lint.gofmt.rules import rules as gofmt_rules
 from pants.backend.go.target_types import GoBinary, GoPackage
 from pants.core.goals.fmt import FmtResult
 from pants.core.goals.lint import LintResult, LintResults
@@ -25,7 +26,7 @@ def rule_runner() -> RuleRunner:
         rules=[
             *external_tool.rules(),
             *fmt.rules(),
-            *gofmt.rules(),
+            *gofmt_rules(),
             *source_files.rules(),
             QueryRule(LintResults, (GofmtRequest,)),
             QueryRule(FmtResult, (GofmtRequest,)),
